calling it a day. Nothing works :D

+++ Server/Entities/Message.cs
@@ -4,8 +4,8 @@ namespace Client.Database.Entities
 {
     public class Message
     {
-        public ulong Id;
-        public ulong AuthorId;
+        public int Id;
+        public int AuthorId;
         public string Text;
         public DateTime Timestamp;
     }
